+++ CHANGELOG.md
@@ -42,6 +42,8 @@
 * add method `Bitrix24\SDK\Core\Credentials\AccessToken::initFromWorkflowRequest`
 * add `\Bitrix24\SDK\Infrastructure\Filesystem\Base64Encoder` for work with base64 encoding
 * add `\Bitrix24\SDK\Core\Exceptions\FileNotFoundException` if file not found
+* add `IncomingRobotRequest` wrapper for data from crm-robot request
+* add `IncomingWorkflowRequest` wrapper for data from biz proc activity request
 * add [Rector](https://github.com/rectorphp/rector) for improve code quality and speed up releases cycle
 
 ## 2.0-beta.2 — 1.04.2024